Broke helper functionality out of test for better visibility.
